+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configurable initial values in method to return the nth Fibonacci number.
 - Method to find the modular multiplicative inverse of a number.
 - Method to calculate n choose k.
+- Method to calculate σ_x(n) (the divisor function).
 
 ## [0.1.2] - 2019-04-29
 ### Added